Use test_helpers module
